migrate sendtx revoke to pom
